requirements.txt
anyio==4.6.2.post1
dlib==19.24.6
face_recognition
fastapi[standard]
Jinja2==3.1.4
numpy==2.1.3
opencv-python==4.10.0.84
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
pandas
python-multipart
httpx